+++ resources/js/v8/components/footers/GalleryFooter.vue
@@ -1,6 +1,8 @@
 <template>
 	<UFooter v-if="footerData" id="footer">
 		<div class="w-full flex flex-col justify-end flex-wrap align-bottom self-end text-center py-5 px-0 text-3xs">
+			<ContactForm v-if="isContactFormVisible" class="w-full text-left mt-4 text-base" />
+
 			<div v-if="footerData.footer_show_social_media" id="home_socials" class="w-full text-muted text-base space-x-2">
 				<a
 					v-if="footerData.sm_facebook_url !== ''"
@@ -54,14 +56,14 @@
 				class="personal_text w-full text-muted leading-6 font-normal"
 				v-html="footerData.footer_additional_text"
 			></p>
+			<p v-if="is_contact_form_enabled && !isContactFormVisible" class="contact_form_link w-full uppercase text-muted leading-6 font-normal">
+				<a rel="noopener noreferrer" target="_blank" :href="Constants.BASE_URL + '/contact'" class="underline">
+					{{ footerData.contact_header ? footerData.contact_header : $t("contact.title") }}
+				</a>
+			</p>
 			<p class="hosted_by w-full uppercase text-muted leading-6 font-normal" v-if="!is_white_label_enabled">
 				<a rel="noopener noreferrer" target="_blank" href="https://lycheeorg.dev" tabindex="-1" class="underline">
 					{{ $t("landing.Powered_by_Lychee") }}
-				</a>
-			</p>
-			<p v-if="isContactLinkVisible" class="contact_form_link w-full uppercase text-muted leading-6 font-normal">
-				<a rel="noopener noreferrer" target="_blank" :href="Constants.BASE_URL + '/contact'" class="underline">
-					{{ footerData.contact_header ? footerData.contact_header : $t("contact.title") }}
 				</a>
 			</p>
 		</div>
@@ -69,6 +71,7 @@
 </template>
 <script setup lang="ts">
 import InitService from "@/services/init-service";
+import ContactForm from "@/v8/components/forms/contact/ContactForm.vue";
 import { useLycheeStateStore } from "@/stores/LycheeState";
 import { storeToRefs } from "pinia";
 import { computed, ref } from "vue";
@@ -82,7 +85,7 @@ const lycheeStore = useLycheeStateStore();
 const { is_white_label_enabled, is_contact_form_enabled, is_contact_form_enabled_on_gallery, is_contact_form_enabled_on_album } =
 	storeToRefs(lycheeStore);
 
-const isContactLinkVisible = computed(() => {
+const isContactFormVisible = computed(() => {
 	if (props.context === "gallery") {
 		return is_contact_form_enabled_on_gallery.value;
 	}
@@ -90,7 +93,7 @@ const isContactLinkVisible = computed(() => {
 		return is_contact_form_enabled_on_album.value;
 	}
 
-	return is_contact_form_enabled.value;
+	return false;
 });
 
 const footerData = ref<App.Http.Resources.GalleryConfigs.FooterConfig | undefined>(undefined);
